User-Service: cart query apis created

+++ User-Service/src/main/java/com/menzo/User_Service/Cart/Controller/CartQueryRestController.java
@@ -71,7 +71,7 @@ public class CartQueryRestController {
     public ResponseEntity<?> getCartItemCount(@RequestHeader("email") String userEmail) {
 
         //  getting cart-item count
-        Integer cartItemCount = cartQueryService.getCartItemCount(userEmail);
+        Long cartItemCount = cartQueryService.getCartItemCount(userEmail);
 
         //  building response
         return ResponseEntity.ok(cartItemCount);
